+++ tests/end2end/test_port.py
@@ -21,7 +21,6 @@ adapters and modify their ports.
 
 from __future__ import absolute_import, print_function
 
-import random
 import warnings
 import pytest
 from requests.packages import urllib3
@@ -32,7 +31,8 @@ from zhmcclient.testutils import hmc_definition, hmc_session  # noqa: F401, E501
 from zhmcclient.testutils import dpm_mode_cpcs  # noqa: F401, E501
 # pylint: enable=line-too-long,unused-import
 
-from .utils import runtest_find_list, TEST_PREFIX, End2endTestWarning
+from .utils import pick_test_resources, runtest_find_list, TEST_PREFIX, \
+    End2endTestWarning
 
 urllib3.disable_warnings()
 
@@ -56,11 +56,10 @@ def test_port_find_list(dpm_mode_cpcs):  # noqa: F811
 
     for cpc in dpm_mode_cpcs:
         assert cpc.dpm_enabled
-        print("Testing on CPC {} (DPM mode)".format(cpc.name))
 
         session = cpc.manager.session
 
-        # Pick a random port of a random adapter
+        # Pick the ports to test with
         adapter_port_tuples = []
         adapter_list = cpc.adapters.list()
         for adapter in adapter_list:
@@ -71,11 +70,14 @@ def test_port_find_list(dpm_mode_cpcs):  # noqa: F811
             msg_txt = "No adapters with ports on CPC {}".format(cpc.name)
             warnings.warn(msg_txt, End2endTestWarning)
             pytest.skip(msg_txt)
-        adapter, port = random.choice(adapter_port_tuples)
+        adapter_port_tuples = pick_test_resources(adapter_port_tuples)
 
-        runtest_find_list(
-            session, adapter.ports, port.name, 'name', 'element-uri',
-            PORT_VOLATILE_PROPS, PORT_MINIMAL_PROPS, PORT_LIST_PROPS)
+        for adapter, port in adapter_port_tuples:
+            print("Testing on CPC {} with port {} of adapter {}".
+                  format(cpc.name, port.name, adapter.name))
+            runtest_find_list(
+                session, adapter.ports, port.name, 'name', 'element-uri',
+                PORT_VOLATILE_PROPS, PORT_MINIMAL_PROPS, PORT_LIST_PROPS)
 
 
 def test_port_update(dpm_mode_cpcs):  # noqa: F811
@@ -88,7 +90,7 @@ def test_port_update(dpm_mode_cpcs):  # noqa: F811
 
     for cpc in dpm_mode_cpcs:
         assert cpc.dpm_enabled
-        print("Testing on CPC {} (DPM mode)".format(cpc.name))
+        print("Testing on CPC {}".format(cpc.name))
 
         adapter_name = TEST_PREFIX + ' test_adapter_crud adapter1'